% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ers.R
\name{adaptive.weights}
\alias{adaptive.weights}
\title{Adaptive Weights}
\usage{
adaptive.weights(coef, sd, n)
}
\description{
Adaptive Weights
}
\examples{
coef = c(0.001, 0.003, -0.004)
sd = c(0.1, 0.4, 0.2)
n = 80
adaptive.weights(coef, sd, n)
}
