% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw_ers_txt_file.R
\name{clean_raw_ers_txt_file}
\alias{clean_raw_ers_txt_file}
\title{Clean a raw ERS text file}
\usage{
clean_raw_ers_txt_file(file, type = c("ERSA", "ERSS", "ERSD", "ERSFA"))
}
\arguments{
\item{file}{path to the raw text file}

\item{type}{ERS file type}
}
\value{
a data frame
}
\description{
Clean a raw ERS text file
}
