% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat}
\alias{dat}
\title{Complete set of example data}
\format{
An object of class \code{data.frame} with 1883 rows and 57 columns.

An object of class \code{data.frame} with 1883 rows and 57 columns.

An object of class \code{data.frame} with 1883 rows and 57 columns.
}
\usage{
dat

dat

dat
}
\description{
Includes the response body mass index at waist, covariates, and log environmental
contaminants.

Includes the response body mass index at waist, covariates, and log environmental
contaminants.

Includes the response body mass index at waist, covariates, and log environmental
contaminants.
}
\keyword{datasets}
