% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ers.R
\name{ers.enet}
\alias{ers.enet}
\title{ERS ENET}
\usage{
ers.enet(
  x,
  y,
  lambda2,
  nfolds = 5,
  foldid,
  pf = rep(1, p),
  pf2 = rep(1, p),
  method = "ls"
)
}
\description{
ERS ENET
}
\examples{
n = length(Y)
nfold = 5
foldid = matrix(data = c(sample(n), rep(1:nfold, length = n)), nrow = n, ncol = 2)
foldid = foldid[order(foldid[,1]),]
p = ncol(metal)
ers.enet(x = metal, y = as.numeric(Y), foldid = foldid[,2], lambda2 = seq(0.00001, 0.01, by = 0.001))
}
