% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ers.R
\name{ers}
\alias{ers}
\title{Environmental Risk Score}
\usage{
ers(
  x,
  y,
  covar = NULL,
  control = list(lambda2.start = NULL, lambda2.adapt = NULL),
  method = "ls",
  scaled = FALSE,
  nfold = 5,
  seed = NULL,
  ...
)
}
\description{
Estimate environmental risk scores from a number of potentially correlated risk
factors using adaptive elastic net (AENET) regression. ENET regression does variable
selection and can select multiple non-zero collinear variables without overfitting.
AENET is an adaptive version of elastic net (ENET) that satisfies the asymptotic
normality assumption that allows us to conduct statistical inference any hypothesis
testing by providing large sample standard errors and p-values.
}
\examples{
set.seed(7794)
fit = ers(x = metal, y = as.numeric(Y), covar = covs,
          control = list(lambda2.start = seq(0.001, 0.5, by = 0.01),
                         lambda2.adapt = seq(0.001, 0.5, by = 0.01)))
}
