% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ersd}
\alias{ersd}
\title{ERSD layout}
\format{
A data frame with 53 rows and 4 variables:
\describe{
\item{Field Name}{field name}
\item{Length}{character length of field}
\item{From}{beginning index}
\item{To}{ending index}
...
}
}
\source{
\url{https://asd.calstate.edu/ded}
}
\usage{
ersd
}
\description{
ERSD layout
}
\keyword{datasets}
