% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ers.R
\name{predict.ers}
\alias{predict.ers}
\title{Make Environmental Risk Score Prediction}
\usage{
\method{predict}{ers}(object, new.x, new.covar = NULL, type = c("class", "link"), ...)
}
\description{
Given an ers fit, predict Environmental Risk Score (ERS)
}
\examples{

}
