% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ers.R
\name{test}
\alias{test}
\title{Test ERS results using performance measures}
\usage{
test(x, y, covar = NULL, fit)
}
\description{
This function explores optimal tuning parameter values for the adaptive
elastic net model. Using the trained model results, environmental risk
score is computed.
}
