% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ers.R
\name{train}
\alias{train}
\title{Train Adaptive ENET Model for ERS}
\usage{
train(x, y, covar = NULL, control = list(method = "cv"))
}
\description{
This function explores optimal tuning parameter values for the adaptive
elastic net model. Using the trained model results, environmental risk
score is computed.
}
\details{
\bold{Algorithm:}
\preformatted{
define sets of model parameter values to evaluate
|for each parameter set
| |for each resampling iteration
| |    hold out specific samples
| |    fit the model on the remainder
| |    predict the hold-out samples
| |  end
|   calculate the average performance across hold-out predictions
|  end
 determine the optimal parameter set
 fit the final model to all the training data using the optimal parameter set
}
}
\examples{

}
