% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y}
\alias{Y}
\title{Example response data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1883 rows and 1 columns.
}
\usage{
Y
}
\description{
A single column with response data. In this case obesity measurement, which
is defined as log body mass index measured at the waist.
}
\details{
\itemize{
\item{Y }{Continuous regression outcome of interest}
}
}
\keyword{datasets}
