% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ers.R
\name{ers}
\alias{ers}
\title{Environmental Risk Score}
\usage{
ers(
  x,
  y,
  covar,
  lambda2.start = NULL,
  lambda2.adapt = NULL,
  method = "ls",
  scaled = FALSE,
  nfold = 5,
  seed = NULL,
  ...
)
}
\description{
Compute environmental risk score
}
\examples{
covar = covs
ers(x = metals, y = as.numeric(Y), covar = covs, 
    lambda2.start = seq(0.00001, 0.01, by = 0.001),
    lambda2.adapt = seq(0.00001, 0.01, by = 0.001))
}
